requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
billiard==4.2.4
boto3==1.42.24
botocore==1.42.24
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
factory_boy==3.3.3
Faker==40.1.0
fastapi==0.128.0
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
kombu==5.6.2
MarkupSafe==3.0.3
orjson==3.11.5
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
structlog==25.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
